separated mechanism_matching.py into a whole subpackage since it was getting too big
